pan_scm_ansible_collection/requirements.txt
pan-scm-sdk>=0.3.0